requirements.txt
nose
sphinx
biopython
editdistance
statistics
matplotlib
numpy
scipy
Pillow
bitarray
dnastorage
